fixing redis bugs
