+++ config.toml
@@ -43,24 +43,28 @@ social_db_contract_id = "v1.social08.testnet"
 # Uncoment the network you want to use or add your own
 
 # mainnet
+# network = "mainnet"
 # rpc_url = "https://rpc.mainnet.near.org"
 # wallet_url = "https://wallet.mainnet.near.org"
 # explorer_transaction_url = "https://explorer.mainnet.near.org/transactions/"
 # rpc_api_key = ""
 
 # testnet
+network = "testnet"
 rpc_url = "https://archival-rpc.testnet.near.org"
 wallet_url = "https://wallet.testnet.near.org"
 explorer_transaction_url = "https://explorer.testnet.near.org/transactions/"
 rpc_api_key = ""
 
 # betanet
+# network = "betanet"
 # rpc_url = "https://rpc.betanet.near.org"
 # wallet_url = "https://wallet.betanet.near.org"
 # explorer_transaction_url = "https://explorer.betanet.near.org/transactions/"
 # rpc_api_key = ""
 
 # shardnet
+# network = "shardnet"
 # rpc_url = "https://rpc.shardnet.near.org"
 # wallet_url = "https://wallet.shardnet.near.org"
 # explorer_transaction_url = "https://explorer.shardnet.near.org/transactions/"
