+++ examples/configs/whitelist_senders.toml
@@ -34,12 +34,14 @@ use_shared_storage = false
 # Uncoment the network you want to use or add your own
 
 # mainnet
+# network = "mainnet"
 # rpc_url = "https://rpc.mainnet.near.org"
 # wallet_url = "https://wallet.mainnet.near.org"
 # explorer_transaction_url = "https://explorer.mainnet.near.org/transactions/"
 # rpc_api_key = ""
 
 # testnet
+network = "testnet"
 rpc_url = "https://archival-rpc.testnet.near.org"
 wallet_url = "https://wallet.testnet.near.org"
 explorer_transaction_url = "https://explorer.testnet.near.org/transactions/"
